tasks/flyover/app/requirements.txt
aiohttp
aiohttp_jinja2
pillow
segno
pypng
